extract ClientTupleSerializer
