+++ modules/client/src/main/java/org/apache/ignite/internal/client/table/ClientRecordBinaryView.java
@@ -39,6 +39,9 @@ public class ClientRecordBinaryView implements RecordView<Tuple> {
     /** Underlying table. */
     private final ClientTable tbl;
 
+    /** Tuple serializer. */
+    private final ClientTupleSerializer ser;
+
     /**
      * Constructor.
      *
@@ -48,6 +51,7 @@ public class ClientRecordBinaryView implements RecordView<Tuple> {
         assert tbl != null;
 
         this.tbl = tbl;
+        ser = new ClientTupleSerializer(tbl.tableId());
     }
 
     /** {@inheritDoc} */
@@ -64,8 +68,8 @@ public class ClientRecordBinaryView implements RecordView<Tuple> {
         return tbl.doSchemaOutInOpAsync(
                 ClientOp.TUPLE_GET,
                 tx,
-                (s, w) -> tbl.writeTuple(tx, keyRec, s, w, true),
-                (s, r) -> ClientTable.readValueTuple(s, r, keyRec));
+                (s, w) -> ser.writeTuple(tx, keyRec, s, w, true),
+                (s, r) -> ClientTupleSerializer.readValueTuple(s, r, keyRec));
     }
 
     /** {@inheritDoc} */
@@ -82,8 +86,8 @@ public class ClientRecordBinaryView implements RecordView<Tuple> {
         return tbl.doSchemaOutInOpAsync(
                 ClientOp.TUPLE_GET_ALL,
                 tx,
-                (s, w) -> tbl.writeTuples(tx, keyRecs, s, w, true),
-                tbl::readTuplesNullable,
+                (s, w) -> ser.writeTuples(tx, keyRecs, s, w, true),
+                ClientTupleSerializer::readTuplesNullable,
                 Collections.emptyList());
     }
 
@@ -102,7 +106,7 @@ public class ClientRecordBinaryView implements RecordView<Tuple> {
         return tbl.doSchemaOutOpAsync(
                 ClientOp.TUPLE_UPSERT,
                 tx,
-                (s, w) -> tbl.writeTuple(tx, rec, s, w),
+                (s, w) -> ser.writeTuple(tx, rec, s, w),
                 r -> null);
     }
 
@@ -120,7 +124,7 @@ public class ClientRecordBinaryView implements RecordView<Tuple> {
         return tbl.doSchemaOutOpAsync(
                 ClientOp.TUPLE_UPSERT_ALL,
                 tx,
-                (s, w) -> tbl.writeTuples(tx, recs, s, w, false),
+                (s, w) -> ser.writeTuples(tx, recs, s, w, false),
                 r -> null);
     }
 
@@ -138,8 +142,8 @@ public class ClientRecordBinaryView implements RecordView<Tuple> {
         return tbl.doSchemaOutInOpAsync(
                 ClientOp.TUPLE_GET_AND_UPSERT,
                 tx,
-                (s, w) -> tbl.writeTuple(tx, rec, s, w, false),
-                (s, r) -> ClientTable.readValueTuple(s, r, rec));
+                (s, w) -> ser.writeTuple(tx, rec, s, w, false),
+                (s, r) -> ClientTupleSerializer.readValueTuple(s, r, rec));
     }
 
     /** {@inheritDoc} */
@@ -156,7 +160,7 @@ public class ClientRecordBinaryView implements RecordView<Tuple> {
         return tbl.doSchemaOutOpAsync(
                 ClientOp.TUPLE_INSERT,
                 tx,
-                (s, w) -> tbl.writeTuple(tx, rec, s, w, false),
+                (s, w) -> ser.writeTuple(tx, rec, s, w, false),
                 ClientMessageUnpacker::unpackBoolean);
     }
 
@@ -174,8 +178,8 @@ public class ClientRecordBinaryView implements RecordView<Tuple> {
         return tbl.doSchemaOutInOpAsync(
                 ClientOp.TUPLE_INSERT_ALL,
                 tx,
-                (s, w) -> tbl.writeTuples(tx, recs, s, w, false),
-                tbl::readTuples,
+                (s, w) -> ser.writeTuples(tx, recs, s, w, false),
+                ClientTupleSerializer::readTuples,
                 Collections.emptyList());
     }
 
@@ -199,7 +203,7 @@ public class ClientRecordBinaryView implements RecordView<Tuple> {
         return tbl.doSchemaOutOpAsync(
                 ClientOp.TUPLE_REPLACE,
                 tx,
-                (s, w) -> tbl.writeTuple(tx, rec, s, w, false),
+                (s, w) -> ser.writeTuple(tx, rec, s, w, false),
                 ClientMessageUnpacker::unpackBoolean);
     }
 
@@ -213,8 +217,8 @@ public class ClientRecordBinaryView implements RecordView<Tuple> {
                 ClientOp.TUPLE_REPLACE_EXACT,
                 tx,
                 (s, w) -> {
-                    tbl.writeTuple(tx, oldRec, s, w, false, false);
-                    tbl.writeTuple(tx, newRec, s, w, false, true);
+                    ser.writeTuple(tx, oldRec, s, w, false, false);
+                    ser.writeTuple(tx, newRec, s, w, false, true);
                 },
                 ClientMessageUnpacker::unpackBoolean);
     }
@@ -233,8 +237,8 @@ public class ClientRecordBinaryView implements RecordView<Tuple> {
         return tbl.doSchemaOutInOpAsync(
                 ClientOp.TUPLE_GET_AND_REPLACE,
                 tx,
-                (s, w) -> tbl.writeTuple(tx, rec, s, w, false),
-                (s, r) -> ClientTable.readValueTuple(s, r, rec));
+                (s, w) -> ser.writeTuple(tx, rec, s, w, false),
+                (s, r) -> ClientTupleSerializer.readValueTuple(s, r, rec));
     }
 
     /** {@inheritDoc} */
@@ -251,7 +255,7 @@ public class ClientRecordBinaryView implements RecordView<Tuple> {
         return tbl.doSchemaOutOpAsync(
                 ClientOp.TUPLE_DELETE,
                 tx,
-                (s, w) -> tbl.writeTuple(tx, keyRec, s, w, true),
+                (s, w) -> ser.writeTuple(tx, keyRec, s, w, true),
                 ClientMessageUnpacker::unpackBoolean);
     }
 
@@ -269,7 +273,7 @@ public class ClientRecordBinaryView implements RecordView<Tuple> {
         return tbl.doSchemaOutOpAsync(
                 ClientOp.TUPLE_DELETE_EXACT,
                 tx,
-                (s, w) -> tbl.writeTuple(tx, rec, s, w, false),
+                (s, w) -> ser.writeTuple(tx, rec, s, w, false),
                 ClientMessageUnpacker::unpackBoolean);
     }
 
@@ -287,8 +291,8 @@ public class ClientRecordBinaryView implements RecordView<Tuple> {
         return tbl.doSchemaOutInOpAsync(
                 ClientOp.TUPLE_GET_AND_DELETE,
                 tx,
-                (s, w) -> tbl.writeTuple(tx, keyRec, s, w, true),
-                (s, r) -> ClientTable.readValueTuple(s, r, keyRec));
+                (s, w) -> ser.writeTuple(tx, keyRec, s, w, true),
+                (s, r) -> ClientTupleSerializer.readValueTuple(s, r, keyRec));
     }
 
     /** {@inheritDoc} */
@@ -305,8 +309,8 @@ public class ClientRecordBinaryView implements RecordView<Tuple> {
         return tbl.doSchemaOutInOpAsync(
                 ClientOp.TUPLE_DELETE_ALL,
                 tx,
-                (s, w) -> tbl.writeTuples(tx, keyRecs, s, w, true),
-                (s, r) -> tbl.readTuples(s, r, true),
+                (s, w) -> ser.writeTuples(tx, keyRecs, s, w, true),
+                (s, r) -> ClientTupleSerializer.readTuples(s, r, true),
                 Collections.emptyList());
     }
 
@@ -324,8 +328,8 @@ public class ClientRecordBinaryView implements RecordView<Tuple> {
         return tbl.doSchemaOutInOpAsync(
                 ClientOp.TUPLE_DELETE_ALL_EXACT,
                 tx,
-                (s, w) -> tbl.writeTuples(tx, recs, s, w, false),
-                tbl::readTuples,
+                (s, w) -> ser.writeTuples(tx, recs, s, w, false),
+                ClientTupleSerializer::readTuples,
                 Collections.emptyList());
     }
